exclude getBytes cross checks
