+++ ci/container/test_component.sh
@@ -76,7 +76,6 @@ for c in "${CATEGORY[@]}"; do
         pushd TestOnly >& /dev/null
             JDBC_VERSION=$(mvn org.apache.maven.plugins:maven-help-plugin:2.1.1:evaluate -Dexpression=project.version --batch-mode | grep -v "[INFO]")
             echo "[INFO] Run JDBC $JDBC_VERSION tests"
-            set -x
             mvn -DjenkinsIT \
                 -Djava.io.tmpdir=$WORKSPACE \
                 -Djacoco.skip.instrument=false \
